Backend v0.11/Techjaysproject/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clickupython==1.0.4
colorama==0.4.6
coverage==4.5.1
coveralls==1.3.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
docopt==0.6.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.6
jwcrypto==1.5.0
more-itertools==10.1.0
oauthlib==3.2.2
pendulum==2.1.2
pluggy==1.3.0
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.8.2
PyJWT==2.8.0
pytest==3.7.1
pytest-cov==2.5.1
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
pytzdata==2020.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
timefhuman==0.0.5
typing-extensions==3.10.0.2
tzdata==2023.3
urllib3==2.1.0
word2number==1.1
wrapt==1.16.0
pandas